requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
gradio==3.50.0
Pillow==9.5.0
typing_extensions==4.5.0
filelock==3.13.1
fsspec==2024.2.0
jinja2==3.1.3
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
sympy==1.13.1
requests
urllib3
pyparsing>=2.4.7
idna==3.10
certifi==2024.12.14
setuptools==75.8.0
anyio==4.6.2.post1
sniffio==1.3.1
pytz==2024.2
pypika==0.48.9
iso8601==2.1.0
dictdiffer==0.9.0
asyncclick==8.1.8
h11==0.14.0
cycler
kiwisolver>=1.4.4
python-dateutil==2.9.0.post0
charset_normalizer==2.0.4
six==1.16.0
billiard==4.2.1
prompt_toolkit==3.0.48
wcwidth==0.2.13
pandas==2.2.3
threadpoolctl
scipy>=1.9.0
joblib